Working ability to make raw ADC reads

+++ include/EVT/io/platform/f3xx/f302x8/ADCf302x8.hpp
@@ -56,7 +56,7 @@ private:
     static Pin channels[MAX_CHANNELS];
     /// Buffer for DMA where each spot represents the value read in from a
     /// channel
-    static uint32_t buffer[MAX_CHANNELS];
+    static uint16_t buffer[MAX_CHANNELS];
     static DMA_HandleTypeDef halDMA;
 
     /**
